Move import_parser into it's own file
