+++ src/python/pants/backend/python/dependency_inference/BUILD
@@ -1,6 +1,12 @@
 # Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-python_sources()
+python_sources(
+    overrides={
+        "import_parser.py": {
+            "dependencies": ["//src/python/pants/backend/python/dependency_inference/scripts:resources"]
+        }
+    }
+)
 
 python_tests(name="tests")
