+++ src/python/pants/backend/python/dependency_inference/scripts/BUILD
@@ -1,0 +1,4 @@
+resources(name="resources", sources=["*.py"])
+
+# Also expose scripts as `python_sources` so they get formatted/linted/checked
+python_sources()
